Begin working on rewrite
The plan is to switch to API only (maybe CSV support can be added back
later) to be able to use statements from all accounts immediately and
be able to auto-detect cross-card statements as well as to clean up
the code by switching to using module instead of single file to be
able to split things up more cleanly. Also remove output to file, it
is handy but just stdout output is more than enough for me.