Make src directory be module named mono2ledger
This begins splitting the code between different files to hopefully
make things more maintainable more in long term by defining config
module for parsing config, cli for outputing to cli and making it so
that main module can be imported on it's own by __init__.py rather
than being used as executable.

Note that config split was done in past so this commit is a bit messy
and has some incomplete parts, but better to start committing changes
now than later.