Simplify cross-card statement detection and printing algorithm
The idea is to use the fact that transactions are ordered properly to
reduce the amount of backtracking that is done with previous
find_source method by changing it to simple iteration. With this
most of methods from classes can be moved to functions as well.

I want to switch to using indexed loop which deletes elements from the
list to be safe from cases where there is transaction and withdraw at
the wrong times which current algo will not detect and probably
mangle result due to use of iterator, but I want to keep the general
workflow detection as is defined here and prefer it over previous
find_source method.