Finish printing
Things are printed in correct order but some statements are missing in
output I guess because of using iterator and some statements before or
after withdraw are ignored. Need to fix this and after I expect there
will only be cleanup and addition of new features if ever needed after
missing transactions are fixed.