SQL cache implementation
This works, but only in case there is monobank transaction on the day
fetch is performed. Otherwise cache will never be used because only
statement timestamp is stored

There are many ways that this can be expanded further like using both
cache & fetch which would make things easier for case where there are
more than 500 statements per period by only fetching non-cached
statements but again

Doing both of these things is not a priority for me as I only use
mono2ledger like once a week and don't have too many transactions, but
will still publish this branch (if not forget) in case this will be
worked on in future.